requirements.txt
python-docx==1.1.2
PyYAML==6.0.2
fastapi==0.115.12
uvicorn==0.34.0
python-multipart==0.0.20
